+++ packages/elements/src/components/ui/action-menu/ActionDropdown.tsx
@@ -78,7 +78,7 @@ export const ActionDropdown: React.FC<ActionDropdownProps> = ({
         ev.preventDefault();
       }
     },
-    [open, closeAndRefocus, expanded]
+    [open, closeAndRefocus]
   );
 
   const contextValue = useMemo(