+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Next release
 
 * Add `ActionDropdown` component to `elements`.
+* Add `useElementFocus` hook to `core`.
 * Add `useMouseIsEntered` hook to `core`.
 
 #### 0.0.35